+++ crates/client/src/rpc/types/cfx/mod.rs
@@ -3,7 +3,6 @@ pub mod account;
 pub mod address;
 pub mod blame_info;
 pub mod block;
-pub mod call_request;
 pub mod consensus_graph_states;
 pub mod epoch_number;
 mod fee_history;
@@ -19,14 +18,15 @@ pub mod storage_collateral_info;
 pub mod sync_graph_states;
 pub mod token_supply_info;
 pub mod transaction;
+pub mod transaction_request;
 pub mod tx_pool;
 pub mod vote_params_info;
 
 pub use access_list::*;
 pub use account::Account;
 pub use address::{
-    check_rpc_address_network, RcpAddressNetworkInconsistent, RpcAddress,
-    UnexpectedRpcAddressNetwork,
+    check_rpc_address_network, check_two_rpc_address_network_match,
+    RcpAddressNetworkInconsistent, RpcAddress, UnexpectedRpcAddressNetwork,
 };
 pub use fee_history::CfxFeeHistory;
 pub use sponsor_info::SponsorInfo;
